refactor: split jetbrainsBase into focused modules
Extract config, gestures, status monitoring, UI traversal and
NVDA compatibility helpers into dedicated modules for better
maintainability and separation of concerns.
